+++ Runtime/ZapparCamera.cs
@@ -150,7 +150,10 @@ namespace Zappar
 
         public void RegisterCameraListener(ICameraListener listener)
         {
-            listeners.Add(listener);
+            if (!listeners.Contains(listener))
+            {
+                listeners.Add(listener);
+            }
         }
 
         private void UpdatePose()
@@ -200,15 +203,9 @@ namespace Zappar
             return m_cameraPose;
         }
 
-        public bool IsMirrored()
-        {
-            return m_isMirrored;
-        }
+        public bool IsMirrored => m_isMirrored;
 
-        public IntPtr GetPipeline()
-        {
-            return m_pipeline;
-        }
+        public IntPtr GetPipeline => m_pipeline;
 
     }
 }